+++ react/tic-tac-toe/src/app/screens/Game/components/Topbar/layout.js
@@ -14,7 +14,7 @@ function TopBar({ handleProfile, handleLogout, userEmail, isLogged }) {
         <Fragment>
           <button className={styles.user} type="button" onClick={handleProfile}>
             <img className={styles.icons} src={userIcon} alt="userIcon" />
-            <span>{`${userEmail} is logged`}</span>
+            <span className={styles.user}>{`${userEmail} is logged`}</span>
           </button>
           <button className={styles.buttonLogOut} type="button" onClick={handleLogout}>
             <span>Log Out</span>
